+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.91.0 (Unreleased)
+BUG FIXES:
+* mongodb: fix migrate from  `resources` to `resources_*` without changes
 
 ## 0.90.0 (May 4, 2023)
 FEATURES: